Implement async metrics in the async debate handler
